apps/gq/src/app/process-case-view/tabs/overview-tab/components/approval-cockpit/approval-workflow-history/approval-workflow-history-iterations/pipes/filter-start-release-events.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';
import {
  ApprovalEventType,
  ApprovalWorkflowEvent,
} from '@gq/shared/models/approval';
@Pipe({
  name: 'filterStartRelease',
})
export class FilterStartReleaseEventsPipe implements PipeTransform {
  transform(events: ApprovalWorkflowEvent[]) {
    if (!events) {
      return [];
    }

    return events.filter(
      (item: ApprovalWorkflowEvent) =>
        item.event !== ApprovalEventType.STARTED &&
        item.event !== ApprovalEventType.RELEASED
    );
  }
}
